backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
ffmpeg-python